[step2] refactor: 중복 코드 제거 및 패키지 이동
